test: expand form refactor coverage
Bring the expanded characterization tests into the stacked refactor branch, keep the refactor-only validation-before-hCaptcha assertion, and move getFormStatusError out of the component file to keep React Doctor clean.
